misalignment structure
