+++ ross/defects/__init__.py
@@ -1,2 +1,1 @@
-from .misalignment_flex import *
-from .misalignment_rigid import *
+from .misalignment import *
